Move getCurrentPrice() method to Quote class
